+++ src/Classifiers/MultilayerPerceptron.php
@@ -404,17 +404,22 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
         ])->check();
 
         if ($this->logger) {
-            $this->logger->info("$this initialized");
+            $this->logger->info("Training $this");
         }
 
         [$testing, $training] = $dataset->stratifiedSplit($this->holdOut);
 
-        [$min, $max] = $this->metric->range()->list();
+        [$minScore, $maxScore] = $this->metric->range()->list();
 
-        $bestScore = $min;
-        $bestEpoch = $delta = 0;
+        $bestScore = $minScore;
+        $bestEpoch = $numWorseEpochs = 0;
         $snapshot = null;
         $prevLoss = INF;
+
+        if ($testing->empty() and $this->logger) {
+            $this->logger->notice('Insufficient validation data, '
+                . 'some features are disabled');
+        }
 
         $this->scores = $this->losses = [];
 
@@ -422,23 +427,24 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
             $batches = $training->randomize()->batch($this->batchSize);
 
             $loss = 0.0;
-            $score = null;
 
             foreach ($batches as $batch) {
                 $loss += $this->network->roundtrip($batch);
             }
 
+            $loss /= count($batches);
+
+            $lossChange = abs($prevLoss - $loss);
+
+            $this->losses[$epoch] = $loss;
+
             if (is_nan($loss)) {
                 if ($this->logger) {
-                    $this->logger->info('Numerical instability detected');
+                    $this->logger->warning('Numerical instability detected');
                 }
 
                 break;
             }
-
-            $loss /= count($batches);
-
-            $this->losses[$epoch] = $loss;
 
             if (!$testing->empty()) {
                 $predictions = $this->predict($testing);
@@ -449,12 +455,18 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
             }
 
             if ($this->logger) {
-                $this->logger->info("Epoch $epoch - {$this->metric}: "
-                    . ($score ?? 'n/a') . ", {$this->costFn}: $loss");
+                $lossDirection = $loss < $prevLoss ? '↓' : '↑';
+
+                $message = "Epoch: $epoch, "
+                    . "{$this->costFn}: $loss, "
+                    . "Loss Change: {$lossDirection}{$lossChange}, "
+                    . "{$this->metric}: " . ($score ?? 'N/A');
+
+                $this->logger->info($message);
             }
 
             if (isset($score)) {
-                if ($score >= $max) {
+                if ($score >= $maxScore) {
                     break;
                 }
 
@@ -464,17 +476,17 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
 
                     $snapshot = Snapshot::take($this->network);
 
-                    $delta = 0;
+                    $numWorseEpochs = 0;
                 } else {
-                    ++$delta;
+                    ++$numWorseEpochs;
                 }
 
-                if ($delta >= $this->window) {
+                if ($numWorseEpochs >= $this->window) {
                     break;
                 }
             }
 
-            if (abs($prevLoss - $loss) < $this->minChange) {
+            if ($lossChange < $this->minChange) {
                 break;
             }
 
@@ -485,7 +497,7 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
             $snapshot->restore();
 
             if ($this->logger) {
-                $this->logger->info("Network restored from snapshot at epoch $bestEpoch");
+                $this->logger->info("Model state restored to epoch $bestEpoch");
             }
         }